requirements.txt
paho_mqtt==1.5.0
